Creating new type to parse the new format of LLM response on task list
Tasks generation can ask the user for more clarification if it is vague or lacks
some actionable and specific information to take an action.
This requires changes to the way we deserialize the tasklist. Now task can potentially be of following format

{
  tasks: [],
  ask_user: "This needs more clarification."
}

Or

{
  tasks: { tasks here },
  ask_user: ""
}

Either tasks or ask_user would be present, not both of them at the same time.