Adding a delay in code-understanding service
Code search servie takes time to start since it needs to load the model first, during this time,
the code-understanding fails to discover the service.

In kuberenetes we can use the health check endpoint which should solve.
But in docker-compose adding just the service dependency is not sufficient to solve this.
Hence adding a delay and retry if code-understanding fails to discover code search.